+++ src/engine/tinanta.rs
@@ -628,6 +628,14 @@ mod tests {
         assert!(f.iter().any(|x| x == "dvezwi"), "{:?}", f);
         let f = generate_all("duha", "plat", 1, 1);
         assert!(f.iter().any(|x| x == "dogDi"), "{:?}", f);
+        let f = generate_all("duha", "alat", 1, 1);
+        assert!(f.iter().any(|x| x == "dugDe"), "{:?}", f);
+        let f = generate_all("duha", "alat", 1, 2);
+        assert!(f.iter().any(|x| x == "duhAte"), "{:?}", f);
+        let f = generate_all("duha", "alat", 1, 3);
+        assert!(f.iter().any(|x| x == "duhate"), "{:?}", f);
+        let f = generate_all("duha", "alrt", 1, 1);
+        assert!(f.iter().any(|x| x == "Dokzyate"), "{:?}", f);
         let f = generate_all("yu", "plat", 1, 1);
         assert!(f.iter().any(|x| x == "yOti"), "{:?}", f);
         let f = generate_all("yA", "plat", 1, 1);